Add ability to configure a proxy for S3 requests
Add an http_options field to the aws_config record. This list of
options is set for the http client prior to making a request and may
include proxy information. Expose versions of the new and configure
functions in the erlcloud_s3 module that allow the proxy host and port
to be set.