Roughing out initial group description call.
I'm using erlcloud_ec2 as a basic example to follow at present and
trying to stay relatively close to what was done there.  Since
the autoscaling group record I've defined will likely need to be
operated on directly outside of this module, I'll probably move
it to an erlang_as.hrl file in the include folder shortly.