Some simple support for paged results from erlcloud_as:describe_groups
I'm not sure I really like the current way it's done but it's a
reasonable start.  I suspect describe_groups/4 may be overkill
or slightly ambiguous but am unsure if simply exposing a parameter
for a list as in the private describe_groups/3 is better from
an API perspective.  Opinions welcome.