fix signature calculation on empty params
when generating a http query string on params with empty lists (ie. queue name prefix on list_queues) the equal sign was not being added next to the key name leaving the empty value (ie. QueueNamePrefix = ). This leads to a signature calculation mismatch since amazon performs the calculation of their side using the equal sign.