erlcloud_xml:get_bool now treats "1" as true.
The API for the IAM get_account_summary specifies that booleans can be 1
or 0.  This caused issue in Inquisitor whereby false was being presented
when in fact the value was true for AccountMFAEnabled.

http://docs.aws.amazon.com/cli/latest/reference/iam/get-account-summary.html
"AccountMFAEnabled This value is 1 if the AWS account (root) has an MFA
device assigned, otherwise it is 0."