Integrate metadata credentials into `default_config/0`
Try metadata credentials as well when `default_config/0` is unable
to resolve redentials from environment or profile. This should allow
access a working config no matter what environment code is running.
Update the expiration field based on the metadata expiration
information.

In addition, utilize the `AWS_SECURITY_TOKEN` if it is set in the
environment (e.g. when running in lambda environment).