erlcloud_ddb_util: Support typed output in get_/q_/scan_all
This change adds an extra option `typed_out :: boolean()` for the multi_call getters in erlcloud_ddb_util.

This meta-option has the effect of setting the `out` option to either `record` or `typed_record`, depending on setting. Currently, the fixed `out` option is `record`, which returns untyped items.  The record output format is used to ensure that the correct response metadata is available to the batching logic in the module, so this is required for the library to work. However, returning the typing data is totally opaque to the functions, and being able to get typed items can help tremendously with resolving ambiguities in edge cases of attributes.