[sqs-batch-entry-spec-error]: Fixup batch_entry() type.
It appears that the batch_entry() type has been doubly wrapped in a list
type, as all usages of the type are also wrapped in a list. From the
naming, I assume the intent was for the type to be singular.

Found while attempting to type check usage of
erlcloud_sqs:delete_message_batch, but problem appears to be general to
all usages of erlcloud_sqs:batch_entry().