Add support for sending and receiving message attributes in SQS
Based on the
[SendMessage](http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_SendMessage.html)
and [ReceiveMessage](http://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_ReceiveMessage.html)
documentation we have added support for including MessageAttributes. Currently, we have
added support for Binary, String, and Number (e.g. integer or
float). The specification includes support for custom types which we
have chosen to exclude in this implementation for the time being. We
have also included the ability to optionally receive message
attributes. The API works similar to the way the existing attributes
API works. A caller can specify, 'none', 'all' or a list of attributes
to include in the response.

At this time we have not included support for MessageAttributes in
send message batch requests.