Custom datatypes in message attributes
Here we have added support for custom message attribute data types for
SQS messages based on the documentation found
http://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-message-attributes.html#message-attributes-data-types-validation.

We have also encoded Float and Integer types using custom data types
since the SQS API only supports a generic Number type. When we receive
a message with a Number attribute we don't know whether it is expected
to be represented as a float or an integer. Previously, we would parse
the data attribute string value for the presence of a `$.`.