Add environment-level configurability for ec2_metadata_url
In some environments, for example, when you're doing K8s on AWS,
you may want to run the likes of a metadata proxy. In this case,
the default EC2 metadata URI needs to be changed. To do this
as transparently as possible, we pass this as an environment variable.

This introduces a general configuration mechanism for passing
erlang application-specific environment variables as OS environment.