Uniform guard clauses for sqs:send_message
Aviod getting a default record if calling sqs:send_message/3 with
`undefined` as the `AwsConfig`.

The old code led to a very weird error as the `undefined` was then used
for `DelaySeconds` and then the guard on sqs:send_message/5 stops it.

Better to fail-fast as soon as we know the argument is wrong.