erlcloud_ddb2: fix issue with serializing GSI creation when billing_mode = PPR
When billing_mode = pay_per_request, updating the global secondary indexes requires a 3-tuple of `{IndexName, KeySpecification, Projection}`, but this has the same signature as a provisioned throughput update for an index, `{IndexName, RU, WU}`. To address this, we need to adjust the serialization to only internpret GSI updates with RU/WU when they are passed as integers. Otherwise the update should get passed into the new GSI creation serialization path.