Add capability to presign URLs with passed-in host header and date
While processing a sigv4 presigned url request, Bookshelf needs to
be able to determine whether the presigned url is valid, appropriately
signed, etc.

The only way I know to do this is to create a new presigned url
using whatever inputs the original presigned url was created with, and
compare the two presigned urls.  In order to do this, I have to use the
same host header and date the original presigned url was created with.
But it wasn't possible to pass in a host header and date, thus this commit.

Signed-off-by: Lincoln Baker <lbaker@chef.io>

code cleanup src/erlcloud_s3.erl

Signed-off-by: Lincoln Baker <lbaker@chef.io>