VPC endpoint configuration: cache availability zone
Follow up to https://github.com/erlcloud/erlcloud/pull/638 (see [comment](https://github.com/erlcloud/erlcloud/pull/638#discussion_r407576219)).

We have a service that uses erlcloud to generate an AWS configuration for accessing the AWS backend based on HTTP requests. It probably should use a shared cache for those configuration structures, but did not need them originally. When we added a VPC Endpoint configuration, the service immediately became degraded, and testing confirmed that it was because we were calling the instance metadata endpoint in a DOS fashion, causing `erlcloud_ec2_meta:get_instance_metadata/2` to mostly take 5 seconds and return `{error, {socket_error, timeout}}`, breaking the VPC endpoint configuration in addition to making most requests to the service take 5+ seconds.

This seems like a hard edge for at-scale uses. It's very hard to understand this problem if you don't understand how this code works, because it doesn't break per se—it just becomes unbearably slow.

To solve the problem, I propose we cache `ok` results from the AZ lookup permanently in the application environment. It's not possible for this value to change during run-time, so this should be totally fine.

I also added some unit tests for this logic.