erlcloud_sqs: get_queue_attributes improvement, test
Follow-up to https://github.com/erlcloud/erlcloud/pull/713.

When AWS added `SqsManagedSseEnabled`, we were lucky that it was a Boolean value, so it didn't break `erlcloud_sqs:decode_attribute_value/1`, which assumes that all attribute values that aren't specifically known are Boolean or integer values. This change anticipates a future when they add a string attribute value (like `QueueArn` or `KmsMasterKeyId `), so that we don't crash decoding THAT.

Finally, adding some `get_queue_attributes` testing to `erlcloud_sqs_tests` to cover this functionality.